refactor test_mlir.py
