+++ new_tests/test_mlir.py
@@ -1,0 +1,26 @@
+""" Tests pyMLIR on examples that use the Toy dialect. """
+import os
+
+from cancer_frontend.numpy_jit_runner import NumpyRunner
+
+MLIR_INPUTS = [
+    "basic.mlir",
+    "broadcast.mlir",
+]
+
+
+def test_basic():
+    np_runner = NumpyRunner()
+    for input_file in MLIR_INPUTS:
+        print("*****************************************************************")
+        print("testing " + input_file + "\n")
+        np_runner.parse_mlir(os.path.join(os.path.dirname(__file__), input_file))
+        np_runner.pretty_mlir()
+        print("")
+        print("*****************************************************************")
+        print("")
+        print("")
+
+
+if __name__ == "__main__":
+    test_basic()
